vertexai/interactions-api/requirements.txt
google-genai==1.59.0